// src/app/core/http/generated/order-returns/v1/model/returnRequest.ts
/**
 * kart-order-service ReturnRequest API (assumed)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Money } from './money';
import { ReturnRequestLineItem } from './returnRequestLineItem';
import { ReturnRequestStatus } from './returnRequestStatus';


export interface ReturnRequest { 
    returnRequestId: string;
    orderId: string;
    customerId: string;
    requestedAmount: Money;
    lineItems: Array<ReturnRequestLineItem>;
    status: ReturnRequestStatus;
    /**
     * Customer-stated return reason.
     */
    reason: string;
    requestedAt: string;
    reviewedBy?: string | null;
    reviewedAt?: string | null;
    rejectionReason?: string | null;
    version: number;
}
export namespace ReturnRequest {
}